Use symbolic links for settings
