+++ Jenkinsfile
@@ -66,8 +66,7 @@ stage('Unittest') {
                     checkout scm
                     sh("""${loginBash}
                           ${moduleDefinition}
-                          cp reframe/settings.py ${publicSettings}
-                          cp ${cscsSettings} reframe/settings.py
+                          ln -sf ${cscsSettings} reframe/settings.py
                           bash ${reframeDir}/${bashScript} -f ${reframeDir} -i ''""")
                 }
             }
